sarvam_mvp/requirements.txt
requests
python-dotenv
click
rich
